+++ core/model/src/main/java/com/atelbay/money_manager/core/model/TransactionOverride.kt
@@ -3,7 +3,7 @@ package com.atelbay.money_manager.core.model
 import kotlinx.datetime.LocalDate
 
 data class TransactionOverride(
-    val amount: Double? = null,
+    val amount: Long? = null,
     val type: TransactionType? = null,
     val details: String? = null,
     val date: LocalDate? = null,
